+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: kanghyki <kanghyki@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/06 15:51:33 by kanghyki          #+#    #+#             */
-/*   Updated: 2022/06/12 03:50:03 by kanghyki         ###   ########.fr       */
+/*   Updated: 2022/06/12 05:19:09 by kanghyki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -65,21 +65,19 @@ void	print_memory_test(char *dst, int size);
  * #                                                       #
  * #########################################################
  */
-/* src/parser/tokenizer/token.c */
-t_token			*ft_init_token(char *pa_str, enum e_token_type token_type);
-void			ft_add_token_back(t_token **head, t_token *new_token);
+/* src/parser/tokenizer/tokenizer.c */
 void			ft_merge_environment(char **str, char **dst, char **env);
 void			ft_quote(char **str, char **dst, char quote, char **env);
 t_token			*ft_create_token_type_metachar(char **str);
 t_token			*ft_create_token_type_argument(char **input_command, char **environment);
-/* Interface */
-t_token	*ft_tokenizer(char *input_command, char **environment);
-/* tokenizer utils */
+t_token			*ft_tokenizer(char *input_command, char **environment); /* Interface */
+/* tokenizer_utils.c */
 char			*ft_strchr_except_null(const char *str, int c);
 char			*ft_strndup(const char *src, size_t n);
-
-/* tokenizer, parser utils */
-void			ft_merge_string(char **dst, char *src);
+t_token			*ft_init_token(char *str, enum e_token_type token_type);
+void			ft_add_token_back(t_token **head, t_token *new_token);
+/* tokenizer_test.c */
+void			ft_print_token_test(t_token *head);
 /*
  * #########################################################
  * #                                                       #
@@ -88,19 +86,22 @@ void			ft_merge_string(char **dst, char *src);
  * #########################################################
  */
 /* src/parser/parser.c */
-t_argument		*ft_init_argument(void);
-char			**ft_init_pa_argument(t_token *cur_token);
-void			ft_add_argument_back(t_argument **head, t_argument *argument);
 char			*ft_get_token_type_char(enum e_token_type token_type);
-t_token			*ft_add_additional_pipe(t_token *cur_token, char **env);
-void			ft_heredoc(t_token *cur_token, char **env);
 t_token			*ft_read_token(t_token *cur_token, t_argument *argument, int index);
-t_token			*ft_read_token_initial_state(t_token *cur_token, t_argument *argument, int index);
-/* Interface */
-t_argument		*ft_parser(char *input_command, char ***environment);
-/* Clean up */
+t_token			*ft_read_token_init(t_token *cur_token, t_argument *arg, int idx);
+t_argument		*ft_parser(char *input_command, char ***environment); /* Interface */
+/* parser_free_utils.c */
 void			ft_free_token(t_token *token);
 void			ft_free_argument(t_argument *arg);
+/* parser_utils.c */
+t_argument		*ft_init_argument(t_token *cur_token, char ***env);
+char			**ft_init_pa_argument(t_token *cur_token);
+void			ft_add_argument_back(t_argument **head, t_argument *arg);
+/* parser_additional.c */
+t_token			*ft_add_additional_pipe(t_token *cur_token, char **env);
+void			ft_heredoc(t_token *cur_token, char **env);
+/* parser_common_utils.c */
+void			ft_merge_string(char **dst, char *src);
 
 /*
  * #########################################################
