+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: kanghyki <kanghyki@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/07 22:42:17 by kanghyki          #+#    #+#             */
-/*   Updated: 2022/06/12 03:33:38 by kanghyki         ###   ########.fr       */
+/*   Updated: 2022/06/12 05:26:24 by kanghyki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
