+++ src/parser/parser.c
@@ -6,57 +6,11 @@
 /*   By: kanghyki <kanghyki@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/03 15:55:58 by kanghyki          #+#    #+#             */
-/*   Updated: 2022/06/12 04:00:57 by kanghyki         ###   ########.fr       */
+/*   Updated: 2022/06/12 05:22:57 by kanghyki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/minishell.h"
-
-t_argument	*ft_init_argument(void)
-{
-	t_argument	*argument;
-
-	argument = (t_argument *)malloc(sizeof(t_argument));
-	if (argument == 0)
-		return (0);
-	ft_memset(argument, 0, sizeof(t_argument));
-	return (argument);
-}
-
-char	**ft_init_pa_argument(t_token *cur_token)
-{
-	char	**pa_argument;
-	int		i;
-
-	i = 0;
-	while (cur_token->next)
-	{
-		if (cur_token->token_type != ARGUMENT)
-			break ;
-		++i;
-		cur_token = cur_token->next;
-	}
-	pa_argument = (char **)malloc(sizeof(char *) * (i + 1));
-	if (pa_argument == 0)
-		return (0);
-	ft_memset(pa_argument, 0, sizeof(char *) * (i + 1));
-	return (pa_argument);
-}
-
-void	ft_add_argument_back(t_argument **head, t_argument *arg)
-{
-	t_argument	*iter;
-
-	if (*head == 0)
-		*head = arg;
-	else
-	{
-		iter = *head;
-		while (iter->next)
-			iter = iter->next;
-		iter->next = arg;
-	}
-}
 
 char	*ft_get_token_type_char(enum e_token_type token_type)
 {
@@ -72,47 +26,6 @@ char	*ft_get_token_type_char(enum e_token_type token_type)
 		return ("<<");
 	else
 		return ("\\n");
-}
-
-t_token	*ft_add_additional_pipe(t_token *cur_token, char **env)
-{
-	char	*read_line;
-	t_token	*add;
-
-	read_line = readline("> ");
-	add = ft_tokenizer(read_line, env);
-	free(read_line);
-	cur_token->next = add;
-	return (cur_token->next);
-}
-
-void	ft_heredoc(t_token *cur_token, char **env)
-{
-	char	*read_line;
-	char	*new_pa_str;
-	char	*for_free;
-	t_token	*new_token;
-
-	new_pa_str = ft_strdup("");
-	read_line = readline("> ");
-	while (ft_strcmp(cur_token->next->pa_str, read_line) != 0)
-	{
-		ft_merge_string(&new_pa_str, read_line);
-		for_free = new_pa_str;
-		new_pa_str = ft_strjoin(new_pa_str, "\n");
-		free(for_free);
-		read_line = readline("> ");
-	}
-	free(read_line);
-	t_token	*next_next_token = cur_token->next->next;
-	free(cur_token->next->pa_str);
-	free(cur_token->next);
-	new_token = ft_tokenizer(new_pa_str, env);
-	free(new_pa_str);
-	cur_token->next = new_token;
-	while (new_token->next != 0)
-		new_token = new_token->next;
-	new_token->next = next_next_token;
 }
 
 t_token	*ft_read_token(t_token *cur_token, t_argument *out_arg, int index)
@@ -133,7 +46,8 @@ t_token	*ft_read_token(t_token *cur_token, t_argument *out_arg, int index)
 				return (ft_add_additional_pipe(cur_token, *(out_arg->env)));
 			else
 			{
-				printf("minishell: parse error near `%s'\n", ft_get_token_type_char(cur_token->next->token_type));
+				printf("minishell: parse error near `%s'\n", \
+						ft_get_token_type_char(cur_token->next->token_type));
 				return (0);
 			}
 		}
@@ -143,7 +57,7 @@ t_token	*ft_read_token(t_token *cur_token, t_argument *out_arg, int index)
 	}
 }
 
-t_token	*ft_read_token_initial_state(t_token *cur_token, t_argument *out_arg, int index)
+t_token	*ft_read_token_init(t_token *cur_token, t_argument *arg, int idx)
 {
 	if (cur_token->token_type != ARGUMENT)
 	{
@@ -153,12 +67,12 @@ t_token	*ft_read_token_initial_state(t_token *cur_token, t_argument *out_arg, in
 			return (0);
 		}
 		else
-			return (ft_read_token(cur_token, out_arg, -1));
+			return (ft_read_token(cur_token, arg, -1));
 	}
 	else
 	{
-		out_arg->pa_argument[index] = cur_token->pa_str;
-		return (ft_read_token(cur_token->next, out_arg, index + 1));
+		arg->pa_argument[idx] = cur_token->pa_str;
+		return (ft_read_token(cur_token->next, arg, idx + 1));
 	}
 }
 
@@ -176,54 +90,17 @@ t_argument	*ft_parser(char *input_command, char ***environment)
 	cur_token = head_token;
 	while (cur_token->token_type != EOL)
 	{
-		cur_arg = ft_init_argument();
-		cur_arg->pa_argument = ft_init_pa_argument(cur_token);
-		cur_arg->env = environment;
-		cur_token = ft_read_token_initial_state(cur_token, cur_arg, 0);
+		cur_arg = ft_init_argument(cur_token, environment);
+		cur_token = ft_read_token_init(cur_token, cur_arg, 0);
 		if (cur_token == 0)
 		{
 			ft_add_argument_back(&head_arg, cur_arg);
 			ft_free_argument(head_arg);
 			ft_free_token(head_token);
-			head_arg = 0;
 			return (0);
 		}
 		ft_add_argument_back(&head_arg, cur_arg);
 	}
 	ft_free_token(head_token);
 	return (head_arg);
-}
-
-void	ft_free_token(t_token *token)
-{
-	t_token *prev;
-
-	prev = token;
-	while (token != 0)
-	{
-		token = token->next;
-		free(prev);
-		prev = token;
-	}
-}
-
-void	ft_free_argument(t_argument *arg)
-{
-	t_argument	*prev;
-	int			i;
-
-	prev = arg;
-	while (arg != 0)
-	{
-		i = 0;
-		while (arg->pa_argument[i] != 0)
-		{
-			free(arg->pa_argument[i]);
-			++i;
-		}
-		free(arg->pa_argument);
-		arg = arg->next;
-		free(prev);
-		prev = arg;
-	}
 }
